simplified grip mechanism
we redesigned the gripping so that the lifting servo only moves down with the item and the slider servo only moves forward without the item.

To reset the configuration you can set the value of the servo to 0 or 180/ vice versa to make it move at full speed in one direction. 90 means stationary.